Reuse the list.Element during Add().
The existing code was popping off the last key, deleting it, and
allocating a new one. But we can just reuse the list.Element and change
the key that it points to.

This both speeds up the core Add loop, but also drops the
bytes-allocated in half. We still seem to have the same spike once we
grow above 1M entries, and a significant growth around 100k, but the
numbers all around are lower.