After seeing how good it was, use it for LRU as well.
The Intern() testing showed how much memory we could save, and
performance was better as well. So do the same mutation for LRU.
The big change is that we can't 'remove' items from the LRU, but we
never implemented support for that anyway. And this gives us much better
performance and memory consumption for the use case that we care about.